// ui/src/app/model/personaContacto.ts
export class PersonaContacto {
    id:          number;
    dni:         string;
    apellido:    string;
    nombre:      string;
    sexo:        string;
    estado:      boolean;
    contactoId:   number;
    pais:         string;
    provincia:    string;
    departamento: string;
    localidad:    string;
    calle:        string;
    numero:       number;
    telefono1:    string;
    telefono2:    string;
    mail:         string;
}